Modify strict mode guard
