+++ sot/utils/__init__.py
@@ -14,7 +14,6 @@ from .utils import (
     NameGenerator,
     ResumeFnNameFactory,
     Singleton,
-    StrictModeGuard,
     count_if,
     execute_time,
     get_unbound_method,
@@ -61,5 +60,4 @@ __all__ = [
     "list_find_index_by_id",
     "show_trackers",
     "get_unbound_method",
-    "StrictModeGuard",
 ]
