+++ tests/test_case_base.py
@@ -1,4 +1,5 @@
 import contextlib
+import os
 import unittest
 
 import numpy as np
@@ -51,3 +52,13 @@ class TestCaseBase(unittest.TestCase):
         sym_output = symbolic_translate(func)(*inputs)
         paddle_output = func(*inputs)
         self.assert_nest_match(sym_output, paddle_output)
+
+
+@contextlib.contextmanager
+def strict_mode_guard(value):
+    if "STRICT_MODE" not in os.environ:
+        os.environ["STRICT_MODE"] = "0"
+    old_value = os.environ["STRICT_MODE"]
+    os.environ["STRICT_MODE"] = str(value)
+    yield
+    os.environ["STRICT_MODE"] = old_value
